pyproject.toml
[project]
name = "amazon-nwlst-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "asyncio-throttle>=1.0.2",
    "beautifulsoup4>=4.14.3",
    "fastexcel>=0.18.0",
    "google-auth>=2.45.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.2",
    "gspread-asyncio>=2.0.0",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "oauth2client>=4.1.3",
    "openpyxl>=3.1.5",
    "playwright>=1.57.0",
    "polars>=1.36.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "scrapy>=2.14.0",
    "xlsxwriter>=3.2.9",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.36.0",
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
